# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/rawatInap": {
            "get": {
                "description": "Get a list of RawatInap.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RawatInap"
                ],
                "summary": "Get all RawatInap.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RawatInap"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new RawatInap.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RawatInap"
                ],
                "summary": "Create New RawatInap.",
                "parameters": [
                    {
                        "description": "the body to create a new RawatInap",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.rawatInapInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RawatInap"
                        }
                    }
                }
            }
        },
        "/rawatInap/{id}": {
            "get": {
                "description": "Get an RawatInap by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RawatInap"
                ],
                "summary": "Get RawatInap.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RawatInap id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RawatInap"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a RawatInap by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RawatInap"
                ],
                "summary": "Delete one RawatInap.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RawatInap id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update RawatInap by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RawatInap"
                ],
                "summary": "Update RawatInap.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RawatInap id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age rawatInap category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.rawatInapInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RawatInap"
                        }
                    }
                }
            }
        },
...
        "/resep": {
            "get": {
                "description": "Get a list of Resep.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resep"
                ],
                "summary": "Get all Resep.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Resep"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Resep.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resep"
                ],
                "summary": "Create New Resep.",
                "parameters": [
                    {
                        "description": "the body to create a new Resep",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.resepInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resep"
                        }
                    }
                }
            }
        },
        "/resep/{id}": {
            "get": {
                "description": "Get an Resep by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resep"
                ],
                "summary": "Get Resep.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resep id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resep"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Resep by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resep"
                ],
                "summary": "Delete one Resep.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resep id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Resep by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resep"
                ],
                "summary": "Update Resep.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resep id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age resep category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.resepInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resep"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.rawatInapInput": {
            "type": "object",
            "properties": {
                "dokterInapID": {
                    "type": "integer"
                },
                "kamarID": {
                    "type": "string"
                },
                "pasienID": {
                    "type": "integer"
                },
                "perawatID": {
                    "type": "integer"
                },
                "tanggal_keluar": {
                    "type": "string"
                },
                "tanggal_masuk": {
                    "type": "string"
                }
            }
        },
        "controller.resepInput": {
            "type": "object",
            "properties": {
                "diagnosaID": {
                    "type": "integer"
                },
                "dosis": {
                    "type": "string"
                },
                "jumlah": {
                    "type": "integer"
                },
                "obatID": {
                    "type": "integer"
                }
            }
        },
...
        "models.RawatInap": {
            "type": "object",
            "properties": {
                "dokterID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "kamarID": {
                    "type": "string"
                },
                "pasienID": {
                    "type": "integer"
                },
                "perawatID": {
                    "type": "integer"
                },
                "tanggal_keluar": {
                    "type": "string"
                },
                "tanggal_masuk": {
                    "type": "string"
                }
            }
        },
        "models.Resep": {
            "type": "object",
            "properties": {
                "diagnosaID": {
                    "type": "integer"
                },
                "dosis": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jumlah": {
                    "type": "integer"
                },
                "obatID": {
                    "type": "integer"
                }
            }
        },
...
    }
}